EXPERIMENTAL Creature move adjustment's
Mostly based on Meth Biscuit Type III's ROM balance patches.

basic attack moves
 (leaf, punch, scratch, shade, shake, spark, squirt)
 power 100->118
 cost 5->4
These moves were all painfully bad in comparison with the 8 and 10 cost
moves (160 and 200 power, respectively). They honestly might still be a
little weak, but hopefully this makes them feel a little better to use.
Incidentally pinch blows all of these moves out of the water with 150
power for 3 energy, but I'm leaving it as-is because I think Weethan is
allowed to do a little funny.

OHKO moves
 (bramble, crush, devour, drown, petrify, shred, terror, thunder,
 vortex)
 cost 20->16
The OHKO effect is strong, but given that they have only 180 power if it
misses 20 energy felt steep for these.
Notably absent from the above list is Wreck, which got its own buff:
 power 175->180
 cost 20->16
 rating skill->str
The power buff is trivial, and the rating change is mostly because I'm
not sure why it was skill to begin with. Given that only Weebo and
Ormagon get it, and both at high levels, this may not matter at all
anyway.

Conversely, Consume gets a nerf:
 power 160->140
 cost 10->14
Honestly, it's still OP with these numbers given that it has OHKO chance
AND energy drain, but aside from the three "special" creatures that get
it (Megathan, Ormagon, and Chaos Jile) only Ergar has it so it's
probably fine.

status moves
 (coil, goo, lamp, roar)
 cost 8->3
Meth Biscuit made these moves do damage and I'm not going that far just
yet but they need a big buff to make up for the opportunity cost of
taking a turn to not do damage. This might make lamp and roar in
particular too strong, though.

big aoe moves (flock, pack)
 cost 35->18
Their power is big (175) but 35 still feels way too high for these
moves.

Swarm gets a similar adjustment:
 power 175->160
 cost 20->12
This is aimed at giving it an identity as a cheaper, somewhat
lower-power version. It's exclusive to Vellup so I'm OK with it being
strong for its cost (the other 12 cost aoes have 100 power).

Spore also gets an adjustment:
 power 100->118
 cost 15->12
It has a 75% chance to cause plague on top of these, so it's extremely
strong. But as Mushroom Hyren's signature move, it should be.

Syphon (Leaf Hyren's signature) was also badly in need of a buff:
 power 90->118
 cost 30->8
It has energy drain, but honestly I have no idea why it cost 30. Its
damage might still be a little low, but its cost is low to compensate.

Timber (Timber Hyren's signature):
 power 175->240
 cost 20->16
I don't know why a Hyren's signature move was strictly worse than
Furok's second ability. Go figure. Kept its cost relatively high, but
gave it ridiculous power to go with it.

Infect (Hook Wing's signature):
 power 150->140
A small nerf because with the plague chance it felt crazy. But it is a
signature move so I didn't want to make it too weak.

Quake (Cave Hyren, Borgor)
 cost 20->12
Yeah, I know, Borgor doesn't need a buff. But it does friendly fire AND
it's basically a Hyren's signature.

Fortune (Rabbage), Turtle (Parathit)
 cost 8->2
Similar to Weethan and Hook Wing, when a pathetic little creature has a
signature move I want it to feel great to use.

Sponge, Soar
 cost 8->4
Some other positive status moves get a buff, with similar logic to the
negative status moves above.
Note that Shield stays at 8 energy because it's a 20 boost instead of a
10.

Mist (Orathan F)
 cost 15->8
Again, status moves have a big opportunity cost. This one is party wide
so its high cost feels more justified, but it's also a hidden creature's
signature.

Impale (Jore, Furok)
 power 160->200
 cost 10->12
 Gave a 10% chance to terminate
Initially I just buffed it to 200 power because it was a strictly worse
version of Maul. Then I noticed that Furok has both and felt very lost.
Honestly, Furok doesn't need more toys. I might take this one away when
I start adjusting creatures.

Heal
 cost 3->5
It needs a nerf to offset the power it gets from its new skill scaling.